详情界面

+++ fast_framework/src/main/java/com/lqs/fast/fast/utils/AppUtil.java
@@ -5,8 +5,12 @@ import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.net.Uri;
+import android.os.Build;
+import android.support.compat.BuildConfig;
+import android.support.v4.content.FileProvider;
 import android.widget.Toast;
 
+import java.io.File;
 import java.util.List;
 
 /**
@@ -41,12 +45,33 @@ public final class AppUtil {
     }
 
     public static void installApk(Context context, String filePath, boolean isOpen) {
-        Intent i = new Intent(Intent.ACTION_VIEW);
-        i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        i.setDataAndType(Uri.parse("file://" + filePath), "application/vnd.android.package-archive");
-        context.startActivity(i);
-        if (isOpen) {
-            android.os.Process.killProcess(android.os.Process.myPid());
+        Intent intent = new Intent(Intent.ACTION_VIEW);
+//        i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+//        i.setDataAndType(Uri.parse("file://" + filePath), "application/vnd.android.package-archive");
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            Uri contentUri = FileProvider.getUriForFile(context, BuildConfig.APPLICATION_ID + ".fileProvider", new File(filePath));
+            intent.setDataAndType(contentUri, "application/vnd.android.package-archive");
+//        } else {
+            intent.setDataAndType(Uri.parse("file://" + filePath), "application/vnd.android.package-archive");
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+//        }
+            context.startActivity(intent);
+            if (isOpen) {
+                android.os.Process.killProcess(android.os.Process.myPid());
+            }
+//
+//        Intent intent = new Intent(Intent.ACTION_VIEW);
+////判断是否是AndroidN以及更高的版本
+//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+//            intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+//            Uri contentUri = FileProvider.getUriForFile(context, BuildConfig.APPLICATION_ID + ".fileProvider", apkFile);
+//            intent.setDataAndType(contentUri, "application/vnd.android.package-archive");
+//        } else {
+//            intent.setDataAndType(Uri.fromFile(apkFile), "application/vnd.android.package-archive");
+//            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+//        }
+//        startActivity(intent);
         }
     }
